requirements.txt
alembic>=1.4.2
amqp>=2.6.0
billiard>=3.6.3.0
celery>=4.4.6
certifi>=2020.6.20
chardet>=3.0.4
click>=7.1.2
envparse>=0.2.0
Faker>=4.1.1
Flask>=1.1.2
Flask-Migrate>=2.5.3
Flask-SQLAlchemy>=2.4.3
future>=0.18.2
gunicorn>=20.0.4
idna>=2.9
importlib-metadata>=1.7.0
itsdangerous>=1.1.0
Jinja2>=2.11.2
kombu>=4.6.11
Mako>=1.1.3
MarkupSafe>=1.1.1
Pillow>=7.1.2
python-dateutil>=2.8.1
python-editor>=1.0.4
pytz>=2020.1
requests>=2.24.0
six>=1.15.0
SQLAlchemy>=1.3.18
text-unidecode>=1.3
urllib3>=1.25.9
vine>=1.3.0
Werkzeug>=1.0.1
zipp>=3.1.0
